refactor: 3D 이미지 별도 컴포넌트화 #40
